+++ pyagentspec/src/pyagentspec/templating.py
@@ -7,7 +7,8 @@
 """Module containing helper functions used to extract placeholders."""
 
 import re
-from typing import List
+from collections import Counter
+from typing import Any, List
 
 from pyagentspec.property import Property
 
@@ -24,10 +25,10 @@ def get_placeholders_from_string(string_with_placeholders: str) -> List[str]:
     )
 
 
-def get_placeholder_properties_from_string(
-    string_with_placeholders: str,
+def get_placeholder_properties_from_json_object(
+    object: Any,
 ) -> List[Property]:
-    """Get the property descriptions for the placeholder names extracted from a string."""
+    """Get the property descriptions for the placeholder names extracted from a JSON serializable object (with nested inputs)."""
     return [
         Property(
             json_schema={
@@ -35,7 +36,42 @@ def get_placeholder_properties_from_string(
                 "type": "string",
             }
         )
-        for placeholder in get_placeholders_from_string(
-            string_with_placeholders=string_with_placeholders
-        )
+        for placeholder in get_placeholders_from_json_object(object=object)
     ]
+
+
+def get_placeholders_from_json_object(
+    object: Any,
+) -> List[str]:
+    """Retrieve the used variable names from any python object.
+    Recursively traverses dicts, lists, sets, tuples, etc. and collects all templated variables in found strings and byte sequences.
+
+    Parameters
+    ----------
+    object : Any
+        A potentially nested python object (str, bytes, dict, list, set, tuple)
+
+    Returns
+    -------
+    List[str]
+        List of the extracted variable names.
+        Note: this list is flattened and does not follow the structure of the inputted object
+    """
+    if isinstance(object, str):
+        return get_placeholders_from_string(object)
+    elif isinstance(object, bytes):
+        return get_placeholders_from_json_object(object.decode("utf-8", errors="replace"))
+    elif isinstance(object, dict):
+        key_templates = get_placeholders_from_json_object(list(object.keys()))
+        value_templates = get_placeholders_from_json_object(list(object.values()))
+        return list(set(key_templates + value_templates))
+    elif isinstance(object, list) or isinstance(object, set) or isinstance(object, tuple):
+        placeholders = Counter(
+            nested_placeholder
+            for nested_item in object
+            for nested_placeholder in get_placeholders_from_json_object(nested_item)
+        )
+        return list(placeholders)
+    else:
+        # unknown object reached, ignore
+        return []