+++ pyagentspec/tests/validation/io_validation/test_toolnode_iovalidation.py
@@ -6,16 +6,9 @@
 
 import pytest
 
-from pyagentspec.flows.edges.controlflowedge import ControlFlowEdge
-from pyagentspec.flows.edges.dataflowedge import DataFlowEdge
-from pyagentspec.flows.flow import Flow
-from pyagentspec.flows.nodes.endnode import EndNode
-from pyagentspec.flows.nodes.mapnode import MapNode, ReductionMethod
-from pyagentspec.flows.nodes.startnode import StartNode
 from pyagentspec.flows.nodes.toolnode import ToolNode
 from pyagentspec.property import (
     FloatProperty,
-    Property,
 )
 from pyagentspec.tools.servertool import ServerTool