+++ docs/pyagentspec/source/changelog.rst
@@ -98,6 +98,11 @@ New features
 Improvements
 ^^^^^^^^^^^^
 
+* **Extended functionality for data parameter**
+
+  Extended the data parameter in ``RemoteTool`` and ``ApiNode`` from only being a dictionary to any JSON serializable object (including nested objects).
+  Also improved template rendering in the ``RemoteTool`` and ``ApiNode`` for PyAgentSpec adapters.
+
 * **Python 3.14 support**
 
   Introduced support for Python version 3.14.